@@ -1,8 +1,8 @@
 package oscar.gmail.com.causality.answer;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
 import android.util.Log;
 
 import java.util.List;
