@@ -1,11 +1,10 @@
 package oscar.gmail.com.causality.question;
 
 import android.app.Application;
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MediatorLiveData;
-import android.arch.lifecycle.MutableLiveData;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MediatorLiveData;
+import androidx.lifecycle.MutableLiveData;
 import android.os.AsyncTask;
-import android.util.Log;
 
 import java.util.List;
 
@@ -20,7 +19,7 @@ public class QuestionRepository {
 
     private QuestionDao mQuestionDao;
     private MutableLiveData<Integer> insertResult = new MutableLiveData<>();
-    private MutableLiveData<List<Question>> questionList = new MutableLiveData<>();
+    private LiveData<List<Question>> questionList;
 
     QuestionRepository(Application application) {
         AppDatabase db = AppDatabase.getDatabase(application);
@@ -32,7 +31,7 @@ public class QuestionRepository {
         return insertResult;
     }
 
-    public MutableLiveData<List<Question>> getQuestionList() {
+    public LiveData<List<Question>> getQuestionList() {
         return questionList;
     }
 
